+++ src/Buerokratt.Common.UnitTests/CentOps/CentOpsServiceTests.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Buerokratt.Common.UnitTests
+namespace Buerokratt.Common.UnitTests.CentOps
 {
     public class CentOpsServiceTests
     {
